Improvement to data extraction from the serializer
Organize the names of the CSV files correctly and set up the paths
