+++ parameters/supported.txt
@@ -1,0 +1,6 @@
+parameters_AURORA_v02.json
+parameters_IREC_version1.json
+parameters_test.json
+IREC_version_03_OPR.json
+IREC_version_03_OPR(1).json
+Aurora_version_04_OPR.json
